Fix error caused by deprecated check for old compilers
Pull in the latest changes from the upstream formula to fix the following error message from Homebrew:

```
Error: Calling 'fails_with :clang' with 'build' < 600 is disabled! There is no replacement.
Please report this to the sebroeder/sofatutor tap:
  /usr/local/Homebrew/Library/Taps/sebroeder/homebrew-sofatutor/Formula/sphinx.rb:30

If reporting this issue please do so at (not Homebrew/brew or Homebrew/core):
  https://github.com/sebroeder/homebrew-sofatutor/issues
```

Also remove the bottles because they are not build against MySQL 5.7 and would render this custom formula useless.